aws파일 연동 설정

+++ src/main/java/org/zerock/service/board/BoardService.java
@@ -4,6 +4,7 @@ import java.io.File;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -11,6 +12,12 @@ import org.zerock.domain.board.BoardDto;
 import org.zerock.domain.board.PageInfo;
 import org.zerock.mapper.board.BoardMapper;
 import org.zerock.mapper.board.ReplyMapper;
+
+import software.amazon.awssdk.core.sync.RequestBody;
+import software.amazon.awssdk.services.s3.S3Client;
+import software.amazon.awssdk.services.s3.model.DeleteObjectRequest;
+import software.amazon.awssdk.services.s3.model.ObjectCannedACL;
+import software.amazon.awssdk.services.s3.model.PutObjectRequest;
 
 @Service
 @Transactional
@@ -22,6 +29,12 @@ public class BoardService {
 	@Autowired
 	private ReplyMapper replyMapper;
 	
+	@Autowired
+	private S3Client s3Client;
+	
+	@Value("${aws.s3.bucket}")
+	private String bucketName;
+	
 	public int register(BoardDto board, MultipartFile[] files) {
 		// db에 게시물 정보 저장
 		int cnt = boardMapper.insert(board);
@@ -31,24 +44,35 @@ public class BoardService {
 				// db에 파일 정보 저장
 				boardMapper.insertFile(board.getId(), file.getOriginalFilename());
 				
-				// 파일 저장
-				// board id 이름의 새폴더 만들기
-				File folder = new File("C:\\Users\\user\\Desktop\\study\\upload\\prj1\\board\\" + board.getId());
-				folder.mkdirs();
-				
-				File dest = new File(folder, file.getOriginalFilename());
-				
-				try {
-					file.transferTo(dest);
-				} catch (Exception e) {
-					// @Transactional은 RuntimeException에서만 rollback 됨
-					e.printStackTrace();
-					throw new RuntimeException(e);
-				}
+				uploadFile(board.getId(), file);
 			}
 		}
 		
 		return cnt;
+	}
+
+	private void uploadFile(int id, MultipartFile file) {
+		try {
+			//S3에 파일 저장
+			
+			String key = "prj1/board/" + id + "/" + file.getOriginalFilename();
+			
+			//putObjectRequest
+			PutObjectRequest putObjectRequest = PutObjectRequest.builder()
+					.bucket(bucketName)
+					.key(key)
+					.acl(ObjectCannedACL.PUBLIC_READ)
+					.build();
+			
+			//requestBody
+			RequestBody requestBody = RequestBody.fromInputStream(file.getInputStream(), file.getSize());
+			//object(파일) 올리기
+			s3Client.putObject(putObjectRequest, requestBody);
+			
+		} catch (Exception e) {
+			e.printStackTrace();
+			throw new RuntimeException(e);
+		}
 	}
 
 	public List<BoardDto> listBoard(int page, String type, String keyword, PageInfo pageInfo) {
@@ -93,14 +117,13 @@ public class BoardService {
 		
 		// removeFiles 에 있는 파일명으로 
 		
-		for (String fileName : removeFiles) {
-			// 1. File 테이블에서 record 지우기
-			boardMapper.deleteFileByBoardIdAndFileName(boardId, fileName);
-			// 2. 저장소에 실제 파일 지우기
-			String path = "C:\\Users\\user\\Desktop\\study\\upload\\prj1\\board\\" + boardId + "\\" + fileName;
-			File file = new File(path);
-			
-			file.delete();
+		if (removeFiles != null) {
+			for (String fileName : removeFiles) {
+				// 1. File 테이블에서 record 지우기
+				boardMapper.deleteFileByBoardIdAndFileName(boardId, fileName);
+				// 2. 저장소에 실제 파일 지우기
+				deleteFile(boardId, fileName);				
+			}
 		}
 		
 		
@@ -114,18 +137,7 @@ public class BoardService {
 				boardMapper.insertFile(boardId, name);
 				
 				// 저장소에 실제 파일 추가
-				File folder = new File("C:\\Users\\user\\Desktop\\study\\upload\\prj1\\board\\" + board.getId());
-				folder.mkdirs();
-				
-				File dest = new File(folder, name);
-				
-				try {
-					file.transferTo(dest);
-				} catch (Exception e) {
-					// @Transactional은 RuntimeException에서만 rollback 됨
-					e.printStackTrace();
-					throw new RuntimeException(e);
-				}
+				uploadFile(boardId, file);
 			}
 			
 		}
@@ -135,17 +147,17 @@ public class BoardService {
 	}
 
 	public int remove(int id) {
-		// 저장소의 파일 지우기
-		String path = "C:\\Users\\user\\Desktop\\study\\upload\\prj1\\board\\" + id;
-		File folder = new File(path);
+		BoardDto board = boardMapper.select(id);
 		
-		File[] listFiles = folder.listFiles();
+		List<String> fileNames = board.getFileName();
 		
-		for (File file : listFiles) {
-			file.delete();
+		if(fileNames!=null) {
+			for(String fileName : fileNames) {
+				// s3저장소의 파일 지우기
+				deleteFile(id, fileName);				
+			}
 		}
 		
-		folder.delete();
 		
 		// db 파일 records 지우기
 		boardMapper.deleteFileByBoardId(id);
@@ -158,6 +170,15 @@ public class BoardService {
 		
 		// 게시물 지우기
 		return boardMapper.delete(id);
+	}
+
+	private void deleteFile(int id, String fileName) {
+		String key = "prj1/board/" + id + "/" + fileName;
+		DeleteObjectRequest deleteObjectRequest = DeleteObjectRequest.builder()
+				.bucket(bucketName)
+				.key(key)
+				.build();
+		s3Client.deleteObject(deleteObjectRequest);
 	}
 	
 }
